feat: extract ANSI logic into src/lib/ansi.ts
Move STRIP_ANSI_RE, stripAnsi, PALETTE_256, rgbStr, AnsiStyle,
parseAnsiSegments, renderAnsiToHtml out of AppView.tsx into
aloop/cli/dashboard/src/lib/ansi.ts. AppView.tsx re-exports all
symbols to preserve the public API surface used by App.tsx and tests.

Aloop-Agent: build
Aloop-Iteration: 2
Aloop-Session: orchestrator-20260321-172932-issue-153-20260323-133712
